requirements.txt
asgiref==3.2.7
backcall==0.1.0
colorama==0.4.3
decorator==4.4.2
Django==3.0.4
django-extensions==2.2.9
django-webpack-loader==0.6.0
factory-boy==2.12.0
Faker==4.0.3
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
parso==0.7.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.5
pydotplus==2.0.2
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
